scsi: Match parameter type to member type.

+++ devices/common/scsi/scsi.h
@@ -256,7 +256,7 @@ public:
         return this->last_selection_message;
     }
 
-    void set_xfer_len(uint64_t len) override {}
+    void set_xfer_len(int len) override {}
 
     void set_status(uint8_t status_code) override {
         this->status = status_code;
