+++ devices/common/scsi/scsihd.h
@@ -57,7 +57,7 @@ protected:
     bool is_device_ready() override { return true; }
 
     // HACK: it shouldn't be here!
-    void set_xfer_len(uint64_t len) override {
+    void set_xfer_len(int len) override {
         this->bytes_out = len;
     }
